upd subscr

+++ backend/recipes/views.py
@@ -73,27 +73,50 @@ class UserViewSet(viewsets.ModelViewSet):
         user.save()
         return Response({"status": "Пароль успешно изменён"}, status=200)
 
-    @action(detail=True, methods=['post'], permission_classes=[permissions.IsAuthenticated])
+    @action(detail=True, methods=['post', 'delete'], permission_classes=[permissions.IsAuthenticated])
     def subscribe(self, request, pk=None):
-        """Подписка на пользователя"""
+        """Подписка или отписка от пользователя."""
         author = get_object_or_404(User, pk=pk)
+
         if request.user == author:
-            return Response({"error": "Нельзя подписаться на самого себя."}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"error": "Нельзя подписаться или отписаться от самого себя."}, status=status.HTTP_400_BAD_REQUEST)
 
-        subscription, created = Subscription.objects.get_or_create(author=author, subscriber=request.user)
-        if created:
-            return Response({"status": f"Вы подписались на {author.username}"}, status=status.HTTP_201_CREATED)
-        return Response({"status": f"Вы уже подписаны на {author.username}"}, status=status.HTTP_400_BAD_REQUEST)
+        # POST запрос - подписка
+        if request.method == 'POST':
+            subscription, created = Subscription.objects.get_or_create(author=author, subscriber=request.user)
+            if created:
+                return Response({"status": f"Вы подписались на {author.username}"}, status=status.HTTP_201_CREATED)
+            return Response({"status": f"Вы уже подписаны на {author.username}"}, status=status.HTTP_400_BAD_REQUEST)
 
-    @action(detail=True, methods=['delete'], permission_classes=[permissions.IsAuthenticated])
-    def unsubscribe(self, request, pk=None):
-        """Отписка от пользователя"""
-        author = get_object_or_404(User, pk=pk)
-        subscription = Subscription.objects.filter(author=author, subscriber=request.user)
-        if subscription.exists():
-            subscription.delete()
-            return Response({"status": f"Вы отписались от {author.username}"}, status=status.HTTP_204_NO_CONTENT)
-        return Response({"error": f"Вы не подписаны на {author.username}"}, status=status.HTTP_400_BAD_REQUEST)
+        # DELETE запрос - отписка
+        elif request.method == 'DELETE':
+            subscription = Subscription.objects.filter(author=author, subscriber=request.user)
+            if subscription.exists():
+                subscription.delete()
+                return Response({"status": f"Вы отписались от {author.username}"}, status=status.HTTP_204_NO_CONTENT)
+            return Response({"error": f"Вы не подписаны на {author.username}"}, status=status.HTTP_400_BAD_REQUEST)
+
+    # @action(detail=True, methods=['post'], permission_classes=[permissions.IsAuthenticated])
+    # def subscribe(self, request, pk=None):
+    #     """Подписка на пользователя"""
+    #     author = get_object_or_404(User, pk=pk)
+    #     if request.user == author:
+    #         return Response({"error": "Нельзя подписаться на самого себя."}, status=status.HTTP_400_BAD_REQUEST)
+
+    #     subscription, created = Subscription.objects.get_or_create(author=author, subscriber=request.user)
+    #     if created:
+    #         return Response({"status": f"Вы подписались на {author.username}"}, status=status.HTTP_201_CREATED)
+    #     return Response({"status": f"Вы уже подписаны на {author.username}"}, status=status.HTTP_400_BAD_REQUEST)
+
+    # @action(detail=True, methods=['delete'], permission_classes=[permissions.IsAuthenticated])
+    # def unsubscribe(self, request, pk=None):
+    #     """Отписка от пользователя"""
+    #     author = get_object_or_404(User, pk=pk)
+    #     subscription = Subscription.objects.filter(author=author, subscriber=request.user)
+    #     if subscription.exists():
+    #         subscription.delete()
+    #         return Response({"status": f"Вы отписались от {author.username}"}, status=status.HTTP_204_NO_CONTENT)
+    #     return Response({"error": f"Вы не подписаны на {author.username}"}, status=status.HTTP_400_BAD_REQUEST)
 
     @action(detail=True, methods=['get'], permission_classes=[permissions.AllowAny])
     def recipes(self, request, pk=None):
